Standardize namespace and fix code scanning errors
