+++ src/AzureIoTHub.Portal.Server.Tests.Unit/Controllers/v1.0/DeviceModelCommandsControllerTests.cs
@@ -68,7 +68,7 @@ namespace AzureIoTHub.Portal.Server.Tests.Unit.Controllers.V10
                 .ReturnsAsync(mockResponse.Object);
 
             _ = this.mockCommandsTableClient.Setup(c => c.Query<TableEntity>(
-                 It.Is<string>(x => x == $"PartitionKey eq '{ entity.RowKey }'"),
+                 It.Is<string>(x => x == $"PartitionKey eq '{entity.RowKey}'"),
                  It.IsAny<int?>(),
                  It.IsAny<IEnumerable<string>>(),
                  It.IsAny<CancellationToken>()))
@@ -323,7 +323,7 @@ namespace AzureIoTHub.Portal.Server.Tests.Unit.Controllers.V10
             var mockResponse = this.mockRepository.Create<Response>();
 
             _ = this.mockCommandsTableClient.Setup(c => c.Query<TableEntity>(
-                It.Is<string>(x => x == $"PartitionKey eq '{ deviceModelId }'"),
+                It.Is<string>(x => x == $"PartitionKey eq '{deviceModelId}'"),
                 It.IsAny<int?>(),
                 It.IsAny<IEnumerable<string>>(),
                 It.IsAny<CancellationToken>()))
