+++ src/AzureIoTHub.Portal.Server.Tests.Unit/Pages/DeviceConfigurations/DeleteDeviceConfigurationTests.cs
@@ -1,7 +1,7 @@
 // Copyright (c) CGI France. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace AzureIoTHub.Portal.Server.Tests.Unit.Pages.LoRaWan.Concentrator
+namespace AzureIoTHub.Portal.Server.Tests.Unit.Pages.DeviceConfigurations
 {
     using System;
     using System.Net.Http;
