+++ src/AzureIoTHub.Portal.Server.Tests.Unit/Pages/DevicesModels/CreateLoRaDeviceModelPageTests.cs
@@ -18,10 +18,8 @@ namespace AzureIoTHub.Portal.Server.Tests.Unit.Pages.DevicesModels
     [TestFixture]
     public class CreateLoRaDeviceModelPageTests : IDisposable
     {
-#pragma warning disable CA2213 // Disposable fields should be disposed
         private Bunit.TestContext testContext;
         private MockHttpMessageHandler mockHttpClient;
-#pragma warning restore CA2213 // Disposable fields should be disposed
 
         private MockRepository mockRepository;
         private Mock<IDialogService> mockDialogService;
