+++ src/AzureIoTHub.Portal/Client/Pages/EdgeDevices/ConnectionStringDialog.razor
