+++ src/AzureIoTHub.Portal/Client/Pages/EdgeDevices/EdgeDeviceDeleteConfirmationDialog.razor
